+++ CompileTimeTracker/Editor/Util/TrackingUtil.cs
@@ -10,8 +10,8 @@ namespace DTCompileTimeTracker {
       return string.Format("{0}s", (ms / 1000.0f).ToString("F2", CultureInfo.InvariantCulture));
     }
 
-    public static long GetMilliseconds() {
-      return DateTime.Now.Ticks / TimeSpan.TicksPerMillisecond;
+    public static int GetMilliseconds() {
+      return (int)(DateTime.Now.Ticks / TimeSpan.TicksPerMillisecond);
     }
   }
 }
